Fixed some warnings under -Wconversion

+++ test/403-read-bandwidth.cc
@@ -34,7 +34,7 @@ SCENARIO("The read queue respects the tag bandwidth") {
     for (long i = 0; i < size; ++i) {
       PACKET seed;
       seed.address = champsim::address{seed_base_addr + i};
-      seed.instr_id = i;
+      seed.instr_id = (uint64_t)i;
       seed.cpu = 0;
 
       seeds.push_back(seed);
@@ -65,7 +65,7 @@ SCENARIO("The read queue respects the tag bandwidth") {
         for (auto elem : elements)
           elem->_operate();
 
-      auto cycle = (size-1)/tag_bandwidth;
+      auto cycle = ((uint64_t)size-1)/tag_bandwidth;
 
       THEN("Packet " + std::to_string(size-1) + " was served in cycle " + std::to_string(cycle)) {
         REQUIRE(mock_ul.packets.back().return_time == mock_ul.packets.back().issue_time + hit_latency + cycle);
@@ -106,7 +106,7 @@ SCENARIO("The prefetch queue respects the tag bandwidth") {
     for (long i = 0; i < size; ++i) {
       PACKET seed;
       seed.address = champsim::address{seed_base_addr + i};
-      seed.instr_id = i;
+      seed.instr_id = (uint64_t)i;
       seed.cpu = 0;
 
       seeds.push_back(seed);
@@ -137,7 +137,7 @@ SCENARIO("The prefetch queue respects the tag bandwidth") {
         for (auto elem : elements)
           elem->_operate();
 
-      auto cycle = (size-1)/tag_bandwidth;
+      auto cycle = ((uint64_t)size-1)/tag_bandwidth;
 
       THEN("Packet " + std::to_string(size-1) + " was served in cycle " + std::to_string(cycle)) {
         REQUIRE(mock_ul.packets.back().return_time == mock_ul.packets.back().issue_time + hit_latency + cycle);
@@ -179,7 +179,7 @@ SCENARIO("The write queue respects the tag bandwidth") {
     for (long i = 0; i < size; ++i) {
       PACKET seed;
       seed.address = champsim::address{seed_base_addr + i};
-      seed.instr_id = i;
+      seed.instr_id = (uint64_t)i;
       seed.cpu = 0;
 
       seeds.push_back(seed);
@@ -210,7 +210,7 @@ SCENARIO("The write queue respects the tag bandwidth") {
         for (auto elem : elements)
           elem->_operate();
 
-      auto cycle = (size-1)/tag_bandwidth;
+      auto cycle = ((uint64_t)size-1)/tag_bandwidth;
 
       THEN("Packet " + std::to_string(size-1) + " was served in cycle " + std::to_string(cycle)) {
         REQUIRE(mock_ul.packets.back().return_time == mock_ul.packets.back().issue_time + hit_latency + cycle);
